refacot: move devloper to localclass and rename file or dart conventions
